+++ src/pages/Dashboard/Payment/index.js
@@ -1,15 +1,15 @@
+import "react-credit-cards/es/styles-compiled.css";
 import { useEffect, useState } from "react";
 import styled from "styled-components";
-import Cards from "react-credit-cards";
-import InputMask from "react-input-mask";
-import "react-credit-cards/es/styles-compiled.css";
+import { toast } from "react-toastify";
+import dayjs from "dayjs";
 
 import useApi from "../../../hooks/useApi";
 
-import SectionSubtitle from "../../../components/Dashboard/NavigationBar/SectionSubtitle";
-import SectionTitle from "../../../components/Dashboard/NavigationBar/SectionTitle";
-import Button from "../../../components/Form/Button";
-import { toast } from "react-toastify";
+import SectionSubtitle from "../../../components/Dashboard/SectionSubtitle";
+import SectionTitle from "../../../components/Dashboard/SectionTitle";
+import ConfirmButton from "../../../components/Dashboard/NavigationBar/ConfirmButton";
+import CreditCard from "../../../components/Dashboard/Payment/CreditCard";
 
 export default function Payment() {
   const [ ticketInfo, setTicketInfo ] = useState(null);
@@ -24,7 +24,7 @@ export default function Payment() {
   const { ticket } = useApi();
 
   useEffect(() => {
-    const { data } = ticket.getTicketInformations();
+    const { data } = ticket.getTicketInformation();
     setTicketInfo(data);
   }, []);
 
@@ -44,7 +44,26 @@ export default function Payment() {
     });
   };
 
+  const isValidCreditCard = () => {
+    if (creditCardInfo.number.replace(/\D/g, "").length < 16)
+      return toast("Número do cartão inválido"); 
+    
+    if (creditCardInfo.name.length < 3)
+      return toast("Nome inválido");
+    
+    if (creditCardInfo.cvc.length < 3)
+      return toast("CVC inválido");
+    
+    if (creditCardInfo.expiry.replace(/\D/g, "").length < 4 || !dayjs(creditCardInfo.expiry).isValid()) {
+      return toast("Validade incorreta");
+    }
+    
+    return true;
+  };
+
   const onSubmitPayment = () => {
+    if (isValidCreditCard() !== true) return;
+
     const ticketInfoData = {
       enrollmentId: 1,
       type: "presential",
@@ -84,122 +103,21 @@ export default function Payment() {
         <p>R$ {ticketInfo?.value}</p>
       </TicketInfo>
       <SectionSubtitle>Pagamento</SectionSubtitle>
-      <CreditCardContainer>
-        <Cards
-          cvc={creditCardInfo.cvc}
-          expiry={creditCardInfo.expiry}
-          focused={creditCardInfo.focus}
-          name={creditCardInfo.name}
-          number={creditCardInfo.number}
-        />
-        <InputsCreditCardContainer>
-          <StyledInputMask
-            type="tel"
-            name="number"
-            placeholder="Número do Cartão"
-            mask="9999 9999 9999 9999"
-            onChange={handleInputChangeCreditCard}
-            onFocus={handleInputFocusCreditCard}
-            required
-          />
-          <StyledInputMask
-            type="text"
-            name="name"
-            placeholder="Nome"
-            maxLength={26}
-            onChange={handleInputChangeCreditCard}
-            onFocus={handleInputFocusCreditCard}
-            required
-          />
-          <RowInputsCreditCard>
-            <StyledInputMask
-              type="tel"
-              name="expiry"
-              placeholder="Validade"
-              mask="99/99"
-              onChange={handleInputChangeCreditCard}
-              onFocus={handleInputFocusCreditCard}
-              required
-            />
-            <StyledInputMask
-              type="tel"
-              name="cvc"
-              placeholder="CVC"
-              maxLength={4}
-              onChange={handleInputChangeCreditCard}
-              onFocus={handleInputFocusCreditCard}
-              required
-            />
-          </RowInputsCreditCard>
-        </InputsCreditCardContainer>
-      </CreditCardContainer>
-      <Button
+      {!ticketInfo.paymentDate ?
+        <CreditCard
+          creditCardInfo={creditCardInfo}
+          handleInputFocusCreditCard={handleInputFocusCreditCard}
+          handleInputChangeCreditCard={handleInputChangeCreditCard}
+        /> : ""
+      }
+      <ConfirmButton
         onClick={onSubmitPayment}
-        disabled={isLoading}
-      >Finalizar Pagamento</Button>
+        isLoading={isLoading}
+      >
+        Finalizar Pagamento
+      </ConfirmButton>
     </>);
 }
-
-const StyledInputMask = styled(InputMask)`
-  border-radius: 5px;
-  padding-left: 15px;
-  font-size: 20px;
-  outline: none;
-  border: solid 1px #C3C3C3;
-
-  @media (max-width: 770px) {
-    font-size: 15px;
-  }
-`;
-
-const RowInputsCreditCard = styled.div`
-  height: 100%;
-  display: flex;
-  justify-content: space-between;
-
-  input {
-    width: 65%;
-  }
-
-  & > input:last-child {
-    width: 30%;
-  }
-`;
-
-const InputsCreditCardContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-left: 15px;
-  padding: 5px;
-
-  & > *:not(:last-child) {
-    margin-bottom: 15px;
-  }
-
-  input {
-    height: 100%;
-  }
-
-  @media (max-width: 770px) {
-    margin-left: 0px;
-  }
-`;
-
-const CreditCardContainer = styled.div`
-  display: flex;
-  max-width: 700px;
-  margin-bottom: 35px;
-
-  @media (max-width: 770px) {
-    flex-direction: column;
-    max-width: 400px;
-
-    & > div:last-child {
-      margin-top: 20px;
-    }
-  }
-`;
 
 const TicketInfo = styled.div`
   background-color: #FFEED2;
